gateways/DARE-CACTUS/requirements.txt
django==1.4.2
django-chunks
django-storages
south==0.7.6
django-celery
django-social-auth